refactor unit tests
Refactor the various unit tests found in semantic.rs into their own
integration tests, exposing the different types as public as needed.
